+++ ci/checks/prohibit-double-backticks-in-python.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -e -o pipefail -u
+
 if git grep -n '``' -- '*.py' ':!docs/generate_py_docs.py'; then
   echo
   echo 'Formatting provided by git_machete.utils.fmt only accepts single backticks (`...`).'